requirements.txt

Flask==3.0.0
Pillow==10.1.0
pillow-heif==0.14.0